Back-port wpt classpath attribute calculation to Gradle core
This commit unifies the way the Tooling API and the eclipse-wtp
plugin defines deployment classpath attributes on an Eclipse project.

The commit aims to fix GRADLE-2123, but also introduces a breaking
change because it changes the content of the Eclipse descriptors generated
by the eclipse and eclipse-wtp plugin. For a complete description check
out https://github.com/gradle/buildship/issues/7.

As a side-effect the Tooling API now considers the manipulations done
in eclipse.project.beforeMerged and eclipse.project.whenMerged closures.

+review REVIEW-6015